+++ core/constants.py
@@ -9,3 +9,14 @@ RICHTEXT_FEATURES__MINIMAL = ()
 
 LESSON_BLOCK = 'lesson'
 PLACEHOLDER_BLOCK = 'placeholder'
+
+AWS_S3_MAIN_HOSTNAME_OPTIONS = [
+    # https://docs.aws.amazon.com/general/latest/gr/s3.html
+    's3.amazonaws.com',  # most likely
+    's3.eu-west-2.amazonaws.com',   # London
+    's3.dualstack.eu-west-2.amazonaws.com',  # London IPv4 + IPv6
+    # 'account-id.s3-control.eu-west-2.amazonaws.com',  # inviable for us
+    # 'account-id.s3-control.dualstack.eu-west-2.amazonaws.com',   # inviable for us
+    's3-accesspoint.eu-west-2.amazonaws.com',
+    's3-accesspoint.dualstack.eu-west-2.amazonaws.com',
+]